refactor(bot): merge command/client type knot to break madge cycle
The last circular dependency (types/CustomClient.ts > models/Command.ts
> types/CommandData.ts) was type-only but still flagged by madge at the
file level. Command's execute signature is parameterized by
CustomClient, and CustomClient's commands/contextMenus fields are
keyed collections of Command/ContextMenuCommand, so splitting them
across files always leaves one file importing back into another.

Move Command, ContextMenuCommand, and CustomClient into
types/CommandData.ts (a single leaf module), and turn
models/Command.ts, models/ContextMenuCommand.ts, and
types/CustomClient.ts into re-export shims so existing import paths
keep working unchanged.

Also fix the madge CI job so it actually blocks: the baseline
tolerated up to 1 cycle (matching the one that just got fixed), so a
reintroduced or new cycle would never trip it once count returned to
1. Baseline is now 0. Also drop the `|| echo 0` fallback on the grep
-c count: grep -c already prints 0 on no match but exits 1, so the
fallback appended a spurious second "0" line and broke the numeric
comparison on the clean-count-of-0 path.

Verified both directions locally: madge exits 0 on the fixed tree and
exits 1 (COUNT=1) when the original three files are restored.

Closes #2330
