+++ packages/bot/src/models/Command.ts
@@ -1,23 +1,6 @@
-import type { CommandCategory } from '../config/constants'
-import type { TCommandData, TCommandExecute } from '../types/CommandData'
-
-type CommandOptions = {
-    data: TCommandData
-    execute: TCommandExecute
-    category: CommandCategory
-    botPermissions?: bigint[]
-}
-
-export default class Command {
-    data: TCommandData
-    execute: TCommandExecute
-    category: CommandCategory
-    botPermissions?: bigint[]
-
-    constructor(options: CommandOptions) {
-        this.data = options.data
-        this.execute = options.execute
-        this.category = options.category
-        this.botPermissions = options.botPermissions
-    }
-}
+// Command now lives in types/CommandData.ts alongside CustomClient and
+// CommandData (they are mutually recursive; see the comment there and
+// decisions/2026-05-16-next-refactor-target-bot-circular-deps.md). This
+// file re-exports it so existing `models/Command` import paths keep
+// working.
+export { Command as default } from '../types/CommandData'
