+++ packages/bot/src/models/ContextMenuCommand.ts
@@ -1,26 +1,7 @@
-import type { CommandCategory } from '../config/constants'
-import type {
-    TContextMenuData,
-    TContextMenuExecute,
-} from '../types/CommandData'
-
-type ContextMenuCommandOptions = {
-    data: TContextMenuData
-    execute: TContextMenuExecute
-    category: CommandCategory
-    botPermissions?: bigint[]
-}
-
-export default class ContextMenuCommand {
-    data: TContextMenuData
-    execute: TContextMenuExecute
-    category: CommandCategory
-    botPermissions?: bigint[]
-
-    constructor(options: ContextMenuCommandOptions) {
-        this.data = options.data
-        this.execute = options.execute
-        this.category = options.category
-        this.botPermissions = options.botPermissions
-    }
-}
+// ContextMenuCommand now lives in types/CommandData.ts alongside
+// CustomClient and CommandData (they are mutually recursive; see the
+// comment there and
+// decisions/2026-05-16-next-refactor-target-bot-circular-deps.md). This
+// file re-exports it so existing `models/ContextMenuCommand` import
+// paths keep working.
+export { ContextMenuCommand as default } from '../types/CommandData'
